requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
djangorestframework-simplejwt==5.2.2
django-filter==23.3
celery==5.2.7
redis==4.4.0
django-mail-templated==2.6.5
pika==1.3.2